LPS-115242 Move to AuditBarChart component
